Change Beer and OrderDetail float properties to double

+++ WholesaleBeer.API/Models/Domain/Beer.cs
@@ -4,8 +4,8 @@
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public float AlcoholContentPercentage { get; set; }
-        public float Price { get; set; }
+        public double AlcoholContentPercentage { get; set; }
+        public double Price { get; set; }
         public Guid BreweryId { get; set; }
 
         // Navigation properties
